Base login server.
Listens to a tcp connection on port 25565 and if found, creates a new ConnectedPlayer object.

That class will then check the message from the client, match the message type with login, and if successful, set the username of the player to the requested string in the package.

A reply will then be sent over the same stream containing both the username and id of the player.